Step one: restructure lobby-related IPC files into client and server subfolders
