+++ src/ipc/lobby/server/login_reply.rs
