+++ src/ipc/lobby/server/nack_reply.rs
